Set santizer settings for all of TileDB (#3843)
This changes ensures that all TileDB source code is compiled and linked
with the required sanitizer options when requested. For the time being
we're living with this hack for running nightly ASAN tests until we
update the build to avoid the need.

This also disables the unit-cppapi-bitsort-filter.cc test when run under
ASAN as it currently fails due to a user-after-free issue. We're
currently unsure on whether we'll be keeping this particular filter so
for now we just avoid the issue when running ASAN. We're keeping the
test around to run in other builds just to prevent unnoticed bitrot.