Don't write fragment metadata incrementally
We were spending a lot of time writing data invidually. This is a quick
experiment to see what sort of speedup we can get by serializing all
data into a single buffer and then writing to disk.

A real implementation would have a growable buffer that would then flush
to disk periodically. For now I was just wanting to see what sort of
speedups a quick hack would give.

A second follow on optimization would be to do all of the GenericTileIO
serialization in parallel.

However, the current improvement is only minimal. On my 2021 16" M1
MacBook Pro a full test run takes roughly 15m. With this patch its down
to 14m. It might be an interesting optimization later on, but for right
not its not a big winner.