Remove bitsort filter.
The bitsort filter implementation was good but the lack of a multi-column filter pipeline made it's integration with the rest of the code a little bit unclean and will lead to code that will be hard to maintain moving forward. This temporarilly removes the integration and the filter until a proper multi-column filter pipeline is in place.

---
TYPE: IMPROVEMENT
DESC: Remove bitsort filter.