Filter pipeline: make chunk range path default and remove other.
When the filter pipeline was changed to filter ranges of chunks, the other path was left around to simplify the work and avoid potential performance impact on the legacy reader. Now that the legacy reader is mostly decommissioned, we can remove the old path safely.

---
TYPE: IMPROVEMENT
DESC: Filter pipeline: make chunk range path default and remove other.