Experimental NAND implementation
A quick experiment to attempt to replace the use of OR with NAND. It
kinda sorta works for non-nullable fields. Nullable negations are not
symmetric so we can't make this a drop in simplification of the query
condition evaluation code. The sparse counting bitmaps also prevent
direct negation so would also not be supportable. Perhaps once that code
is removed we can deal with the nullable negation and apply this
optimization at that point.