Fix buffer size error in deserializing query (#3851)
This bug was originally found in MariaDB, but could be easily reproduced using C/C++ APIs: In the case of incomplete queries the client could potentially error out thinking that the data he received from the REST server for a dimension/attribute could not fit in the user buffers.

We have tracked down the root cause to be that, in the case of Incomplete queries, the client had a wrong idea of the actual size of the users buffers, which could perfectly fit the data received from the server. The reason is that after a query returns from the REST server, and only in the client side, we update the user buffer sizes to the actual data sizes received from server, so that the user can check the result size on reads. So if the next data received from the server are more than the data received in the previous query.submit() the client thinks that they don't fit the user buffers.

The reason this bug was not found earlier in that Python API that is mostly used by customers reallocate user buffers on incomplete queries, so the sizes are always reset to the right size. However, MariaDB uses C/C++ APIs and it was very easy to hit this problem with any var-sized data.

So, in order to work around this bug, we use the original_buffer_sizes that are stored along with the actual buffer_sizes to determine if the data received from the REST server fit, instead of the actual buffer_sizes. This was already being done on the Server side so it was straightforward to extend to the client.

After this fix, a reader bug has also been uncovered, where when no Subarray is set for a sparse unordered with duplicates query, the reader was failing to detect when he was done processing all tiles of a fragment, and it would either hit an assert in debug or get in an infinite loop in release. This PR addresses this issue as well.

TYPE: BUG
DESC: Fix buffer size error in deserializing query

---------

Co-authored-by: Robert Bindar <robertbindar@gmail.com>