Fragment consolidation: using correct buffer weights.
This change corrects buffer weights in the fragment consolidator for var sized attributes. Before this change, the fixed/var buffer weights were mixed. The issue manifested itself as a segfault for nullable attributes as the number of validity cells in the buffers was incorrect.

---
TYPE: IMPROVEMENT
DESC: Fragment consolidation: using correct buffer weights.