TMP: HODGE PODGE: Remove resizes, avoid copies
Some changes I chucked in cause I'm feeling lazy on an experimental
branch. There was a performance bug in `string_at` where I wasn't
passing vectors by reference which caused `~vector()` to be a
significant amount of time in performance analysis.

The removal of the resize calls was me trying to figure out if they were
expensive in some unknown way. Turns out not.