Fix bug when setting var sized dimension buffers
This fixes a bug when checking the number of coordinates on each
dimension. Previously, when a variable sized dimension's data buffer was
set via `set_data_buffer` we were accidentally overwriting
`coords_info_.coord_num_` with zero. Thus, if the last dimension data
buffer set is for a variable sized dimension, the writer would think
there was no data to write.

The root cause here was that by not checking whether the dimension is
variable sized, we calculated the number of coordinates in
`set_data_buffer` as `buffer_size / UINT64_MAX` which is zero for all
values of `utin64_t` except `UINT64_MAX`.

The fix is to not attempt to check the number of coordinates in
`set_data_buffer` for variable sized dimensions and instead rely on the
check in `set_offsets_buffer`.