Dense reader: adjust memory configuration parameters. (#3918)
* Dense reader: adjust memory configuration parameters.

This change removes an unnecessary check in the dense reader that made queries fail on REST servers. At some point in the design of the new memory management for the dense reader, the check made sense but it is no longer required.

It also adjusts the new memory variable for the dense reader name to be more intuitive. The variable is a limit to how much tile data we will load into memory at once before we process what we have and then load more data but it is not a memory budget. If a single tile doesn't fit in that upper limit, we might still allow to load it if it is still possible to load it into memory without busting the `sm.mem.total_budget` considering everything else loaded at the time.

Finally, it moves the added `sm.mem.tile_upper_memory_limit` to experimental as it is only exposed for the moment to allow us to guide users to potentially tweak their queries if we need to but we don't want to commit to support this setting moving forward as a way to optimize queries long term.

---
TYPE: IMPROVEMENT
DESC: Dense reader: adjust memory configuration parameters.