Support in place update of group members (#3928)
This adds support for updating group members. We also modify the on disk format for a v2 group members in group details to facilitate the new updating support. A new API is not added in this PR but the act of removing and adding in the same group open is now supported.

I've also reshuffled the code to add a new GroupDetails object which is a missing layer of abstraction from the previous group implementation. This separates the on-disk format (GroupDetails) from the overarching group class.

---
TYPE: FEATURE
DESC: Support in place update of group members