Simplify the Win32 VFS code. (#3785)
This PR simplifies the Win32 VFS code that reads and writes from files. For file writing we call the `WriteFile` function in only one place, and if it writes less bytes than we told it, we don't fail but continue with the rest. And for reading we use an `OVERLAPPED` to specify the offset and avoid using `SetFilePointerEx` (which [has been described by Microsoft as an anachronism](https://learn.microsoft.com/en-US/archive/blogs/winserverperformance/designing-applications-for-high-performance-part-iii-2).

---
TYPE: IMPROVEMENT
DESC: Refactor and simplify Win32 VFS code.

===

* Simplify the file write loop.

WriteFile is called in only one place, and if we wrote less bytes than we told it to write, we don't fail and continue from there.

* Use an `OVERLAPPED` to set the offset in `ReadFile`.

* Format code.

* Address PR feedback.

* Add a test writing and reading large files.

* Read large files in a loop.

* Increment the buffer pointer in the reading loop.

* Improve the large file test to write a pattern and hash it for verification.

* Give a unique tag to the Windows filesystem test.

* Improve creation and deletion of the temporary directory.