Fixup vcpkg CI caching (#3960)
I noticed that the vcpkg caching was broken after we merged. This work fixes up the two main issues. The first issue was that run-vcpkg's default hash keys were non-unique across our builders. Adding the hashed contents of the CC, CXX, and bootstrap_args to the key fixes this issue.

The second issue is that the patch we're applying to triplet files changes the generated package hashes which means that our caches were never used. We just have to add a bit of conditional logic to prevent multiple applications.

Lastly, this also includes a number of other useful fixes I found while working out the two issues above.

---
TYPE: IMPROVEMENT
DESC: Fix vcpkg caching in CI