Sparse global order reader: deletes and overflow can give wrong results.
This fixes a potential yet very difficult to hit issue in global order reads with no duplicates and deletes where the reader could give wrong results. The issue could arise if we hit a var size overflow while copying data to the user buffers and while reverting progress on the cell slab structure, we revert past a deleted cell with duplicates. In some cases, it would possible to revert the index in the read state for one of the duplicates that wasn't deleted in one fragment but not in the fragment where the cell was deleted. This would cause the earlier cell to show up in the results on the read continuation.

The fix is to add an empty cell slab for a deleted cell when it had non deleted duplicates so that when we revert progress past that cell, we update the index in the read state structure.

---
TYPE: IMPROVEMENT
DESC: Sparse global order reader: deletes and overflow can give wrong results.