Sparse unordered w/ dups: smaller units of work. (#3948)
* Sparse unordered w/ dups: smaller units of work.

This change processes smaller units of work for the sparse unordered with duplicates reader the same way the dense reader does to limit memory movement performance penalties during the unfiltering and copy operations.

Also in this change we get rid of the memory budget ratio for query condition. We already treat query condition tiles as coordinates so we can merge the ratio for coordinates and the one for query condition into one.

---
TYPE: IMPROVEMENT
DESC: Sparse unordered w/ dups: smaller units of work.