Implement typed filter pipelines and TypedViewFilter
This PR adds output typing to Filter objects, to allow filters to
their output and constrain their acceptable input. This functionality
is used to implement the almost-noop TypedViewFilter, which allows an
output to be cast to an acceptable input typed if needed.

The typed pipeline functionality resolves SC-24079, where we currently
throw an error if a DoubleDelta filter is used on an attr/dim with a
floating point datatype even if there is an intermediate type change
via the FloatScaleFilter (which outputs a configurable integer
datatype).

- Implement FilterPipeline::check_filter_types; refactor
- Add tests for FilterPipeline::check_filter_types
- Add TypedViewFilter
- Add tests Filter::output_datatype and specialization for float scale filter
- Update tiledb/sm/filter/typed_view_filter.cc to fix constructor type
  Thanks to Julia Dark <24235303+jp-dark@users.noreply.github.com>
- Add serialization for TypedViewFilter

Fixes SC-26926 SC-24079 SC-24526