Fix path list argument in ExternalProject_Add
Our list of paths argument was being misinterpreted as a set of
arguments to the cmake CLI. Previously, when invoking the child cmake
process it was passing `-DCMAKE_PREFIX_PATH=/path1 /path2 /path3` which
then led to warnings about ignoring the extra paths /path2 and /path3.

This change uses the documented option for instructing
ExternalProject_Add to use a separate list combining character instead
of the default `;` which it uses internally.