Sparse global order reader: defer tile deletion until end of merge. (#4014)
This fixes an issue in the sparse global order reader where an unused tile might be used after deletion. The problem arises when a fragment consolidated with timestamps has many duplicated cells. If the tile doesn't get used in the merge, it gets deleted when we process the last cell, but the problem is that there still might be other cells from the same tile in the tile queue as cells in the queue are not sorted depending on position.

The fix is to keep track of all the tiles to delete and delete them when the merge is done to prevent any use after free.

Note that this PR doesn't add unit tests for the issue as the randomness makes it very difficult to come up with any cases where this reproduces consistently.

---
TYPE: IMPROVEMENT
DESC: Sparse global order reader: defer tile deletion until end of merge.

(cherry picked from commit e86556ff4f2572249d6e6f38f6ee18c57b8925f3)