Make sure metadata are deleted when they got added in the same operation. (#4022)
[SC-25089](https://app.shortcut.com/tiledb-inc/story/25089/creating-mutating-and-deleting-metadata-elements-in-one-transaction-results-in-them-still-being-around)

I added the story's reproduction for groups as a test. It fails before the fix and succeeds after it. Ideally it would reside in the metadata unit tests but I can do it as part of the refactorings in #4012; let's fix the bug now.

---
TYPE: BUG
DESC: Fix array/group metadata deletion when they are inserted in the same operation.