Moving unfiltered data to larger allocations.
This moves the unfiltered data buffer from the Tile class to a larger allocation and provide a way to get a pointer to the data. At the moment, the tile class still stores a pointer into the larger allocation but eventually all accesses will be changed to get the tile data from the UnfilteredData class.

It also modifies the sparse global order reader to only load more tiles from a fragment once it is fully processed.

---
TYPE: IMPROVEMENT
DESC: Moving unfiltered data to larger allocations.