Introduce new option for FilterPipeline to run on entire tile
This adds a new C api `tiledb_filter_list_set_use_tile_chunking` and C++
api `FilterList::set_use_tile_chunking` to let the user toggle the
behavior.

The effects of this are if the chunking (splitting a tiling into 64kib)
is used for applying the filters. If set to true (default), tiles are
chunked for filtering (compression). If set to false the entire tile is
considered and passed into the filter pipeline.

In some cases, such as dictionary encoding or RLE on strings the
behavior is automatically set to use tile level filtering.

The effects of tile level filtering are that compressors and encoders
can consider the entire tile's data. If there is a significant amount of
repeated values or if each cell has a large amount of var length data,
allowing the compressors to work on more data can yield better
compression results.

The trade off of tile level filtering is that if the tile size is large
(recall that the chunking defaults to 64kib) then the
decompression/decoding might take longer per tile and depending on how
many tiles are being used by a query its possible to end up with less
efficient use of CPU resources while bottlenecked on the slowest to
decompress tile.