Aggregates: internals for basic sum/count aggregates for sparse reads.
This implements the internals for basi count/sum aggregates. Internals meaning only the execution objects are implemented and not the API layer. The objects are stored in the query and passed the the readers at strategy creation time. Note that aggregates are only implemented for the default channel (which is the channel currently used for reading data) so the aggregates are only able to aggregate the data over one value for the full data set, no window functions are currently supported. It is possible to request aggregates and the data at the same time. No optimization is done to use the tile sums stored in the fragment metadata at the moment. Also, for now aggregates are also only implemented for the sparse global order and sparse unordered with duplicates readers.

---
TYPE: IMPROVEMENT
DESC: Aggregates: internals for sum/count aggregates for sparse reads.