Fix REST client curl buffer bug
It turns out if the size of the query plust the leading 8 byte length
field ended up matching the size of our scratch buffer exactly, we would
end up processing the same query twice. This lead to the situation where
the client would think its buffers were full which would then lead to
errors.

This ensures that if we process any query message, those bytes are
removed from the REST scratch buffer to ensure progress.