Add new STSProfileWithWebIdentityCredentialsProvider
This new credential provider support chaining of assume from based on a
profile configuration including support for web identity tokens as part
of the chain. This is based on the upstream
`STSProfileCredentialProvider`. The upstream credential provider lacks
support for the web identity tokens.

The main use case for this is inside EC2/ECS/EKS environments where a
web identity token can be used as part of IRSA sequences.