Add new STSProfileWithWebIdentityCredentialsProvider (#4137)
* Add new STSProfileWithWebIdentityCredentialsProvider

This new credential provider support chaining of assume from based on a
profile configuration including support for web identity tokens as part
of the chain. This is based on the upstream
`STSProfileCredentialProvider`. The upstream credential provider lacks
support for the web identity tokens.

The main use case for this is inside EC2/ECS/EKS environments where a
web identity token can be used as part of IRSA sequences.

* Minimize object library linking for s3

* Disable s3 object library

This is a workaround until thr curl interfacr correctly handles system
libraries.

* Update tiledb/sm/filesystem/s3/STSProfileWithWebIdentityCredentialsProvider.cc

* Format

* Remove s3 provider cmake object library

This is a temporary workaround. In a followup we will add this back and
correctly setup the filesystem object library. The filesytem object
library had a few issues with s3 that expanded beyond the scope of this
PR.

* Add support for "vfs.s3.config_source"= "sts_profile_with_web_identity"

---------

Co-authored-by: Isaiah Norton <ihnorton@users.noreply.github.com>
Co-authored-by: Isaiah Norton <isaiah@tiledb.io>