Fix vcpkg dependencies building only in Release mode. (#4159)
[SC-30065](https://app.shortcut.com/tiledb-inc/story/30065/the-custom-vcpkg-ports-cause-the-dependencies-to-always-be-built-in-release-mode)

We were trying to prevent vcpkg from building dependencies in both Debug and Release mode by checking the value of `CMAKE_BUILD_TYPE` and setting `VCPKG_BUILD_TYPE` accordingly.

However this approach is not correct. Besides not working with multi-config generators, the triplet files are evaluated by CMake in script mode where `CMAKE_BUILD_TYPE` is always empty, which causes the dependencies to be built only in Release mode, and cause linker errors when the main library is built in Debug mode.

This PR removes this trick. Local builds will unfortunately have to pay the price of the dependencies being built twice, but CI performance will not regress (on Unix we directly patch the official triplet files and on Windows we don't seem to be using vcpkg).

---
TYPE: BUG
DESC: Fix vcpkg dependencies building only in Release mode.