Improve support for multi-config CMake generators. (#4147)
To set specific compiler options for Release or Debug mode, we have been checking the value of the `CMAKE_BUILD_TYPE` variable.

Howeber that variable applies only to single-file CMake generators (Makefiles and Ninja). We still use it for multi-config generators like MSBuild on Windows with the `-EnableDebug` option, but this is non-standard and invalidates the concept of multi-config since the same compiler options will be applied to all configurations.

We fix this by varying the compiler options with the help of CMake generator expressions like [`$<CONFIG>`](https://cmake.org/cmake/help/latest/manual/cmake-generator-expressions.7.html#configuration-expressions). The `-EnableDebug` option in the Windows bootstrap script also got deprecated; just don't specify it when bootstrapping and you can build for both Debug and Release.

---
TYPE: IMPROVEMENT
DESC: Improve support for multi-config CMake generators.