Remove unnecessary transitive dependencies. (#4190)
[SC-32037](https://app.shortcut.com/tiledb-inc/story/32037/fix-lzma-linkage-in-2-16-1)

We have been linking to liblzma when Azure or S3 are enabled, while actually it is only needed on Azure and on non-Windows. This PR removes this dependency.

Should fix [CI failures in TileDB-Java](https://github.com/TileDB-Inc/TileDB-Java/actions/runs/5589366233/job/15138136419#step:4:118). I have a branch ready to backport to 2.16, will need to run a test build with it in Java.

---
TYPE: BUILD
DESC: Fix static linking to the on Windows.