Legacy reader: reading non written region causes segfault.
This fixes an issue with legacy dense reads where reading a region that is not written causes segfault. In this case, the reader generates a cell slab with a nullptr tile, which signals the copy function to return the fill value for those cells. The nullptr value was causing an issue when trying to determine if a field is present or not.

---
TYPE: IMPROVEMENT
DESC: Legacy reader: reading non written region causes segfault.