Remove stderr debug logging in tests
While the root cause is unclear, it appears that logging to stderr in
Catch2 tests can confuse either Catch2, ctest, or CMake on Windows. I
have a vague recollection that stdio on Windows is significantly
different than on Posix so my guess is that something platform specific
was interacting badly with the Windows build system checks.