Refactor enumeration loading
The previous version of loading enumerations accidentally exposed the
concept of multiple array schemas to users of the C API which was a
first. Even though it was only a boolean flag, it was still crossing a
line in our API model so it has been removed.

This also changes how the ArrayDirectory loads enumerations slightly so
that loading multiple enumerations is done in parallely any time more
than one enumeration is required.