Fix regression test exit status; fix unhandled throw (expected)
tiledb_regression was not returning the runner status, so section failures did
not lead to overall failure status of the runner executable.

Also this masked an unhandled throw in one test, which is otherwise expected to
throw because the corresponding fix was already committed.

---
TYPE: BUG
DESC: Fix regression test exit status; fix unhandled throw (expected)