Backwards compatibility: handle 0 chunk var tiles. * Backwards compatibility: handle 0 chunk var tiles.
In arrays with format version earlier than 10, it was possible to create a var file with zero chunks when all cells had an empty string value. This caused segfaults with later versions of the library and is fixed with this change.

---
TYPE: BUG
DESC: Backwards compatibility: handle 0 chunk var tiles.