Fix compilation with GCC 13 (#4331)
There were a number of warnings issued when compiling with GNU g++13 that all appear to be spurious (which fails our -Werror builds). These fixes avoid the warnings and things compile just fine.

The changes to Subarray were mostly to avoid the use of the new threeway comparison operator on vectors. However, in hindsight, using a std::unodred_map with std::string_view keys should also be a (probably tiny) performance improvement since we're not longer copying vectors around as well as switching to an O(1) contianer instead of std::map's O(log N).

The changes to the unit tests in various places look to be obvious compiler optimization bugs in that the warning most likely is missing that the vector internals will expand the internal buffer size before applying the memcpy warning that was causing overwrite errors.

---
TYPE: BUG
DESC: Fix compilation with GCC 13