Fix and optimize the Release CI workflow. (#4373)
This PR fixes failures in the Release workflow on macOS and arm64. #4348 started using vcpkg to detect the current triplet, but due to a bug it thought we were building for x64-osx. I reverted part of the previous logic for macOS.

Also, I noticed that the Release workflow was building the `ExternalProject`s for some of our dependencies, even while vcpkg was enabled, causing them to be built twice. This happened because we were enabling `TILEDB_FORCE_ALL_DEPS`. I stopped passing the option and marked it as incompatible with `TILEDB_VCPKG`.

---
TYPE: BUILD
DESC: Fix and optimize the Release CI workflow.

* Stop passing `TILEDB_FORCE_BUILD_DEPS` during release.

Combining this option with enabling vcpkg causes the dependencies to build with both vcpkg and ExternalProjects. Disabling it causes only vcpkg to be used, which builds everything either way.

* Disable `TILEDB_FORCE_ALL_DEPS` if `TILEDB_VCPKG` is enabled.

* Set the correct triplet on Apple Silicon.

Apparently vcpkg has trouble detecting it by itself.

* Do not set the triplet if it is already set.