Avoid empty HTTP requests to load Enumerations (#4369)
Previously, we weren't checking if we actually had any enumerations to load which resulted in REST requests for an empty list of enumerations. That just wastes time and resources for no reason so this change avoids the request whn there's no work to do.

---
TYPE: IMPROVEMENT
DESC: Avoid empty Enumeration REST requests.