Prevent out of bounds reads
This test case explains part of a bug in the GlobalOrderWriter. When our
fragment is full at the beginning of a write to an existing fragment, we
would accidentally call `dim_tiles[-1].cell_num()` which would actually
manage to return a reasonable value instead of segfaulting. This was due
to `num_tiles_to_write` returning 0 coupled with our failure to check
for that situation until after a call to set_coords_metadata.

Swapping from the `std::vector::operator[]` to `std::vector::at` adds
bounds checking and now throws an exception instead of returning a
garbage value. This is better since without bounds checking we were
corrupting the last tile of every fragment when setting the wrong last
tile cell num.