Fixup query condition set membership creation API (#4390)
During review there was some confusion over having a default of
`TILEDB_IN` when creating set membership query conditions. The confusion
also prompted me to realize that we may have other types of query
conditions in the future that take a list of values that are not set
membership related which makes having the `TILEDB_IN` default even less
well designed.

Unfortunately, I forgot to remove the default from the std::string
version of that function. @eddelbuettel also kindly pointed out that I
forgot to document the `op` paramter which I've also added.