Remove FragmentMetadata::get_footer_size
This function existed to select which of the various footer size calculations
is needed for calculating the footer size for various older format
versions. As simple as it looked, it's actually an extremely subtle
implementation of that decision.

The reason for this is the selection was based on the result of a call
to utils::parse::get_fragment_name_version. The "name version" here is
important as that function's orignal purpose was to return the version
of a fragment name. I.e., our current __t1_t2_uuid_version format is fragment
name version 3. Name version 2 was __t1_t2_uuid, and name version 1 was
__uuid_timestamp. On a historical note, there were even older variations
of name version 1 that went __ospid_timestamp, __threadid_timestamp,
__macaddressthreadid_timestamp, back to __threadid_timestamp, and then
finally, __uuid_timestamp. None of that matters to us now of course.

Over time, the utils::parse::get_fragment_name_version function evolved
into also returning what ended up being something akin to a "footer version"
when the name version is 3 (again, remember that this is different than the
fragment version). This footer version was then used by get_footer_size
to select the appropriate function to calculate the footer size. The
footer size was basically a synthetic version that represented ranges of
the fragment format version which then mapped to get_footer_size_vX_vY
function names, where X and Y were the fragment format ranges for the
associated footer size calculation.

If you've found all that a touch confusing, that feeling is precisely
the motivation for removing this function. This changes things to just
get the fragment format version and then picks the correct function
directly rather than using the not-actually-specified footer version.