Remove FragmentMetadata::footer_size_v11_or_higher
The removal of get_footer_size in the previous commit makes it obvious
that this function is dead code. The range of possible outputs of
`utils::parse::get_fragment_name_version` meant that this function was
never actually called. By removing that function and adjusting the logic
to use the fragment format version directly, it becomes much more clear
that its dead code as evidenced by the fact that we don't have linker
errors after its removal.