Remove direct access to get_fragment_name_version
This isolates the idea of a fragment name version behind the single
unified format version. For old fragments that did not contain a version
in their name, we return the largest format version represented by that
name format. None of the checks on the fragment name version relied on
being able to differentiate these versions (because how could they).

I'll also note that this uncovered and fixed a couple latent bugs. In
both FragmentMetaConsolidator::consolidate and DeletesAndUpdates::dowork
we were failing to check the fragment version return value of UINT32_MAX
which used to indicate "unknown" versions from very old fragments.