Fix non-vcpkg S3 builds. (#4392)
Regardless of whether we use vcpkg or not to acquire the AWS SDK, _importing_ it in our build involves a call to `find_package(AWSSDK)`. Based on that fact, this PR makes the vcpkg and non-vcpkg cases more similar to each other.

What was the problem? For non-vcpkg builds we were crafting our own CMake imported libraries, and failed to include s2n, a non-Windows dependency. Now we are doing what the vcpkg builds do, and use the targets exported by the AWS SDK. Because VCF builds the Core without vcpkg in its nightly builds, they started failing with linker errors related to s2n.

Fixes TileDB-Inc/TileDB-VCF#583. [A VCF nightly build with this fix was launched and succeeded.](https://github.com/TileDB-Inc/TileDB-VCF/actions/runs/6356027166/job/17265037282)

---
TYPE: NO_HISTORY