Remove extraneous casts to void (#4377)
* Remove extraneous casts to void.

This PR removes a number of extraneous casts to void, though not all in the same way. It does not have a goal of removing all of extraneous casts.

The most frequent by far is to prevent unused variable warnings for function arguments that are not actually used. In this case changing to an anonymous argument solves the problem. It also indicates quite clearly that an argument isn't used. In some cases checks for the validity of unused arguments were also removed.

In some cases a variable declaration (argument or otherwise) were annotated with `[[maybe_unused]]`. This was most often the case where the code varies with preprocessor definitions.

In a few cases some argument checks were promoted from `assert` to throwing exceptions when invalid. A handful of ignored return values were replaced with `std::ignore`.

* Replace uses of `std::ignore =` with void casts.

* gcc asserted its superiority by warning that I didn't need an attribute that other compilers use, so I patted it on the head and told it go play nicely with others.

* The children couldn't play nicely, so I had to take away the ball.

* Revert to original code in `signal_handlers.cc` because void-casting causes warnings in that particular circumstance.

* Restored argument names that were apparently, at one point, unused but are now very much used.
clang-format