Stop building gRPC when GCS is enabled. (#4435)
Backport of microsoft/vcpkg#34569 to version 2.15.1 of the
`google-cloud-cpp` port. The first commit copies the port as-is to the
`ports` directory, and the second applies the fix.

Dramatically drops build times and binary size. Before this PR, the size
of the `build/vcpkg_installed/x64-windows` directory with features
`gcs;tests;webp` enabled was __2.5GB__. With this PR it is __708MB__.

---
TYPE: BUILD
DESC: Stop building gRPC when GCS is enabled.