Query aggregates REST support (#4415)
Add rest support for query aggregates.

Notes to reviewers:
- At this stage we should pay special attention to the capnp
specification, make sure it's as generic as possible according to the
aggregates design document.
- I tested this manually with a local rest server and serialization
paths seem to work as expected, but overall the feature doesn't
currently work because we need Go support on the REST side which I'll
file as followup (When the REST server deserializes the query, the query
buffers are set to `nullptr` and their sizes to the correct sizes passed
by the user. The Go code is supposed to allocate memory and set the
buffers again before the query is submitted "locally". The tests work
because in the serialization wrappers we emulate this allocation with
some stack allocated buffers, see changes in `helpers.cc`)

---
TYPE: NO_HISTORY | FEATURE
DESC: Query aggregates REST support

---------

Co-authored-by: Luc Rancourt <lucrancourt@gmail.com>
Co-authored-by: KiterLuc <67824247+KiterLuc@users.noreply.github.com>