Decouple the code in the `fragment` directory from `StorageManager`. (#4453)
[SC-23374](https://app.shortcut.com/tiledb-inc/story/23374/move-storagemanager-load-fragments-to-array)

The `StorageManager` methods that load fragment info and metadata were
moved to the `FragmentInfo` and `FragmentMetadata` classes. These
classes also use `ContextResources` instead of `StorageManager`.

I was going to make a `fragment` object library but it also depends on
`rest` which is not yet componentized.

---
TYPE: IMPROVEMENT
DESC: Refactor the fragment info loading code to reduce coupling.