Aggregates: allow aggregates pipeline the processing of full tiles. (#4465)
This adds a new aggregate_full_tile interface to iaggregator so that we
can later use the data in the fragment metadata to improve aggregation
performance. The data is packaged in the new FullTileData class, which
will later be produced by the fragment metadata.

Note that this only implements the aggregators side of the work...
Reader/FragmentMetadata changes will follow.

---
TYPE: FEATURE
DESC: Aggregates: allow aggregates pipeline the processing of full
tiles.