Convert C API interface functions to CAPI_INTERFACE (#4460)
Convert C API interface functions to use `CAPI_INTERFACE`. This
completes the work in #4430.

There are a very small number of functions not converted. One is not in
the C API (i.e. it doesn't have `extern "C"` linkage), but is used by
the C++ API, and on top of that it's for a deprecated operation. The
others are small, informational functions that do return values directly
and do not use `capi_return_t`.

---
TYPE: NO_HISTORY
DESC: Convert C API interface functions to CAPI_INTERFACE