Stop including the static library when releasing on Windows. (#4469)
This PR stops passing `-EnableStaticTileDB` on the Windows release
workflow, which means that the release artifacts will include include
only the dynamic library, like on other platforms. The size of these
artifacts on Windows drops from 175MB [to just
7MB](https://dev.azure.com/TileDB-Inc/CI/_build/results?buildId=35970&view=artifacts&pathAsName=false&type=publishedArtifacts).

Users that want to statically link to TileDB will have to build it
themselves. TileDB-Java used to do that until
TileDB-Inc/TileDB-Java#314, and according to [a
search](https://github.com/search?q=TileDB%3A%3Atiledb_static++NOT+is%3Aarchived+NOT+is%3Afork&type=code)
of the `TileDB::tiledb_static` string on public non-archived and
non-forked repositories, none of these uses would be broken.

> [!NOTE]
> TileDB-R links statically on Windows as well but it does not use these
artifacts so it will not be affected.

---
TYPE: BREAKING_BEHAVIOR
DESC: The Windows release artifacts no longer include static libraries.