Remove UnitTestConfig::array_encryption_key_length
This field existed to force invalid key lengths. This is unnecessary
since we can just set the invalid key lengths in the test. The reason I
stopped to remove this is because the test state is improperly persisted
between unit tests. If you run `tiledb_unit -d y [capi][encryption]` it
will fail because the UnitTestConfig is not reset.

It took me so long to debug this issue that I decided to light it on
fire and just remove the whole global state and write the tests
properly.