Use `FetchContent` to get vcpkg. (#4484)
[SC-36205](https://app.shortcut.com/tiledb-inc/story/36205/use-fetchcontent-to-acquire-vcpkg)

Currently, building TileDB from a GitHub archive fails by default[^1],
because we use vcpkg from a submodule, and if we are not on a cloned
repository that submodules does not exist.

With this PR we change our approach to using CMake's
[`FetchContent`](https://cmake.org/cmake/help/latest/module/FetchContent.html)
module, and remove the `externals/vcpkg` submodule. [This is what the
Azure SDK for C++
does.](https://github.com/Azure/azure-sdk-for-cpp/blob/12407e8bfcb9bc1aa43b253c1d0ec93bf795ae3b/cmake-modules/AzureVcpkg.cmake#L14-L40)

This has the disadvantage that by default[^1] the vcpkg repository will
be downloaded for every clean configure. It can by mitigated by
specifying a `VCPKG_ROOT` environment variable to a local path of the
vcpkg repository.

[^1]: Unless the vcpkg toolchain file is manually specified.

---
TYPE: BUILD
DESC: Remove the vcpkg submodule and download it automatically when
configuring