Aggregates: enable full support. (#4489)
This removes the need for a configuration setting to use aggregates. It
will be merged into 2.18.

Also fixed is the serialization code paths so that we can run queries
with aggregates on 2.18. I tested that the aggregates work with a REST
server.

---
TYPE: NO_HISTORY
DESC: Aggregates: enable full support.