Remove function `sanity_check()` for contexts
The function `sanity_check()` for contexts has returned a constant `TILEDB_OK` for a while now; that check has moved into the exception wrappers. At the time, the function was stubbed out to avoid combining the large number of lines that would change with other, more significant one. This PR completes the removal of this function.