Remove function `sanity_check()` for contexts (#4506)
The function `sanity_check()` for contexts has returned a constant
`TILEDB_OK` for a while now; that check was moved into the exception
wrappers. At the time, the function was stubbed out to avoid combining
the large number of lines that would change with other, more significant
ones. This PR completes the removal of this function.

---
TYPE: NO_HISTORY
DESC: Remove function `sanity_check()` for contexts