Wrap scandir in an excepion safe wrapper.
This wraps calls to the scandir posix API in a class so that the objects created by the API are freed on destruction.

The class only exposes an index operator and a size API to get access to the underlying dirent structures. Note that no iterator is provided to enable range based loops as this would overcomplicate the class, which will only be used in a handful of places.

---
TYPE: FEATURE
DESC: Wrap scandir in an excepion safe wrapper.